Mob Programming session on March 29. We created an AddressBook App that manages contacts.
The BookAddress class has instance variables for storing the first name, lastname, phone number, email and age of a contact.
It also has methods for getting and setting these values.

The Service class has an instance variable for storing a list of contact objects. It also contains methods for adding a contact to the list, removing a contact from the list, retrieving a list of all contacts and for searching the list of contacts by firstname and lastname.